Move more types around
