+++ .env
@@ -1,0 +1,9 @@
+# defaults to localhost
+#VITE_SERVER=UT201P18.RCH.STGLABS.IBM.COM
+VITE_SERVER=iopen.iinthecloud.com
+
+# defaults to 8076
+VITE_PORT=22
+
+VITE_DB_USER=SNDBX3
+VITE_DB_PASS=SNDBX3
