+++ package.json
@@ -2955,6 +2955,7 @@
 	"scripts": {
 		"lint": "eslint src/**/*.js --no-error-on-unmatched-pattern",
 		"pretest": "npm run lint",
+		"test": "vitest",
 		"package": "vsce package",
 		"build": "rm -rf dist && tsc",
 		"vscode:prepublish": "webpack --mode production",
@@ -2978,6 +2979,7 @@
 		"ssh2-streams": "^0.4.10",
 		"ts-loader": "^9.3.1",
 		"typescript": "^4.8.2",
+		"vitest": "^2.1.8",
 		"webpack": "^5.94.0",
 		"webpack-cli": "^4.5.0"
 	},
