+++ src/api/Search.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import { GetMemberInfo } from './components/getMemberInfo';
-import { IBMiMember, SearchHit, SearchResults } from '../typings';
+import { IBMiMember, SearchHit, SearchResults } from './types';
 import { Tools } from './Tools';
 import IBMi from './IBMi';
