+++ src/commands/open.ts
@@ -1,15 +1,17 @@
 import { commands, Disposable, l10n, QuickInputButton, QuickPickItem, QuickPickItemButtonEvent, QuickPickItemKind, Range, TextDocument, TextDocumentShowOptions, ThemeIcon, Uri, window } from "vscode";
-import { MemberItem, OpenEditableOptions, WithPath } from "../typings";
+import { MemberItem, QsysFsOptions, WithPath } from "../typings";
 import Instance from "../Instance";
 import { Tools } from "../api/Tools";
 import { getUriFromPath, parseFSOptions } from "../filesystems/qsys/QSysFs";
-import { DefaultOpenMode } from "../config/Configuration";
 import path from "path";
 import { findExistingDocument, findExistingDocumentUri } from "../views/tools";
 import IBMi from "../api/IBMi";
+import { DefaultOpenMode } from "../api/configuration/ConnectionManager";
 
 const CLEAR_RECENT = `$(trash) Clear recently opened`;
 const CLEAR_CACHED = `$(trash) Clear cached`;
+
+export type OpenEditableOptions = QsysFsOptions & { position?: Range };
 
 export function registerOpenCommands(instance: Instance): Disposable[] {
