+++ src/filesystems/local/deployment.ts
@@ -4,11 +4,11 @@ import tar from 'tar';
 import tmp from 'tmp';
 import vscode from 'vscode';
 import { instance } from '../../instantiate';
-import { DeploymentParameters } from '../../typings';
 import IBMi from '../../api/IBMi';
 import { Tools } from '../../api/Tools';
 import { getLocalActions } from './actions';
 import { DeployTools } from './deployTools';
+import { DeploymentParameters } from '../../typings';
 
 export namespace Deployment {
   export interface MD5Entry {
