+++ src/filesystems/local/types.ts
@@ -1,0 +1,3 @@
+import { Ignore } from "ignore"
+import { WorkspaceFolder } from "vscode"
+import { DeploymentMethod } from "../../api/types"
