+++ src/views/debugView.ts
@@ -4,8 +4,8 @@ import { checkClientCertificate, debugKeyFileExists, remoteCertificatesExists } 
 import { DebugConfiguration, getDebugServiceDetails, SERVICE_CERTIFICATE } from "../api/configuration/DebugConfiguration";
 import { DebugJob, getDebugServerJob, getDebugServiceJob, isDebugEngineRunning, readActiveJob, readJVMInfo, startServer, startService, stopServer, stopService } from "../debug/server";
 import { instance } from "../instantiate";
-import { BrowserItem } from "../typings";
 import { withContext } from "./tools";
+import { BrowserItem } from "./types";
 
 const title = "IBM i debugger";
 type Certificates = {
