+++ vitest.config.ts
@@ -1,0 +1,13 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  test: {
+    // ... Specify options here.
+    fileParallelism: false,
+    root: './src/api',
+    setupFiles: [
+      'tests/globalSetup.ts',
+    ],
+  },
+})
